requirements.txt
# Core Dependencies
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.0  # Pinned to 1.0.0 for compatibility with crewai

# AI & ML
crewai==0.28.4
crewai-tools==0.4.4
openai==1.30.5

# LangChain
altair==5.3.0
langchain==0.1.12
langchain-community==0.0.28  # Updated to be compatible with langchain 0.1.12
langchain-core==0.1.52
langchain-openai==0.1.7  # Updated to be compatible with embedchain
langchain-text-splitters==0.0.1
langsmith==0.1.67

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Optional: For async processing
# redis==5.0.1
# celery==5.3.6

# Development
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0

# Monitoring & Instrumentation
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-exporter-otlp-proto-http==1.25.0